+++ R/tibblify-package.R
@@ -3,7 +3,6 @@
 
 ## usethis namespace: start
 #' @import rlang
-#' @import vctrs
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 #' @importFrom rlang arg_match0
@@ -13,21 +12,14 @@
 #' @importFrom rlang check_dots_empty
 #' @importFrom rlang current_call
 #' @importFrom rlang current_env
-#' @importFrom rlang exec
 #' @importFrom rlang is_empty
 #' @importFrom rlang is_named
 #' @importFrom rlang is_true
 #' @importFrom rlang list2
-#' @importFrom vctrs list_sizes
-#' @importFrom vctrs obj_check_vector
-#' @importFrom vctrs vec_any_missing
+# These vctrs imports are used in C code.
+#' @importFrom vctrs list_unchop
 #' @importFrom vctrs vec_cast
-#' @importFrom vctrs vec_check_size
-#' @importFrom vctrs vec_detect_missing
-#' @importFrom vctrs vec_init
-#' @importFrom vctrs vec_ptype
-#' @importFrom vctrs vec_size
-#' @importFrom vctrs vec_unique_loc
+#' @importFrom vctrs vec_is
 #' @useDynLib tibblify, .registration = TRUE
 ## usethis namespace: end
 NULL