+++ tests/testthat/helper-functions.R
@@ -3,14 +3,14 @@ new_difftime <- function(units) {
 }
 
 new_rational <- function(n = integer(), d = integer()) {
-  n <- vec_cast(n, integer())
-  d <- vec_cast(d, integer())
+  n <- vctrs::vec_cast(n, integer())
+  d <- vctrs::vec_cast(d, integer())
 
-  size <- vec_size_common(n, d)
-  n <- vec_recycle(n, size)
-  d <- vec_recycle(d, size)
+  size <- vctrs::vec_size_common(n, d)
+  n <- vctrs::vec_recycle(n, size)
+  d <- vctrs::vec_recycle(d, size)
 
-  new_rcrd(list(n = n, d = d), class = "vctrs_rational")
+  vctrs::new_rcrd(list(n = n, d = d), class = "vctrs_rational")
 }
 
 read_sample_json <- function(x) {